Linked SQL to the user input; New stored procedures
Due to an issue with QueryDef objects (ArcObjects fail) there are now
three stored procedures which are run depending on whether the user
includes a geometry field in the query.

User input now linked to the stored procedure and tested

Group By clause not currently allowed for spatial data

New ArcMapModule updates to add feature and table layers to current map.